+++ mindspore/ops/kernel/gpu/cuda/arrays/triu_gpu_kernel.cc
@@ -59,7 +59,8 @@ int TriuGpuKernelMod::Resize(const std::vector<KernelTensor *> &inputs, const st
   }
   ResetResource();
   std::vector<int64_t> output_shape = outputs.at(kIndex0)->GetShapeVector();
-  size_t output_elements_ = std::accumulate(output_shape.begin(), output_shape.end(), 1, std::multiplies<int64_t>());
+  size_t output_elements_ =
+    std::accumulate(output_shape.begin(), output_shape.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
   if (output_elements_ == 0) {
     is_null_input_ = true;
   }
